updating type names
